# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kali Academy cybersecurity learning platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ModuleDifficulty = typeof ModuleDifficulty[keyof typeof ModuleDifficulty];


export const ModuleDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Module {
  id: number;
  title: string;
  slug: string;
  description: string;
  difficulty: ModuleDifficulty;
  category: string;
  lessonCount: number;
  order: number;
  completedCount?: number;
}

export type ModuleWithLessonsDifficulty = typeof ModuleWithLessonsDifficulty[keyof typeof ModuleWithLessonsDifficulty];


export const ModuleWithLessonsDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Lesson {
  id: number;
  moduleId: number;
  title: string;
  slug: string;
  order: number;
  duration: number;
  completed?: boolean;
}

export interface ModuleWithLessons {
  id: number;
  title: string;
  slug: string;
  description: string;
  difficulty: ModuleWithLessonsDifficulty;
  category: string;
  order: number;
  lessons: Lesson[];
}

export interface Command {
  command: string;
  description: string;
  output: string;
  dangerous?: boolean;
}

export interface LessonDetail {
  id: number;
  moduleId: number;
  title: string;
  slug: string;
  order: number;
  duration: number;
  content: string;
  commands: Command[];
  completed?: boolean;
  /** @nullable */
  nextLessonId?: number | null;
  /** @nullable */
  prevLessonId?: number | null;
}

export interface Progress {
  id: number;
  lessonId: number;
  completedAt: string;
}

export interface ProgressInput {
  lessonId: number;
}

export interface Stats {
  totalLessons: number;
  completedLessons: number;
  totalModules: number;
  completedModules: number;
  totalHours: number;
  currentStreak: number;
  level: string;
  xp: number;
}

export type FeedItemType = typeof FeedItemType[keyof typeof FeedItemType];


export const FeedItemType = {
  tip: 'tip',
  achievement: 'achievement',
  lesson_complete: 'lesson_complete',
  module_complete: 'module_complete',
} as const;

export interface FeedItem {
  id: number;
  type: FeedItemType;
  title: string;
  description: string;
  timestamp: string;
  /** @nullable */
  lessonId?: number | null;
}