# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kali Academy cybersecurity learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules/{id}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a module with its lessons
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module with lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleWithLessons"
        "404":
          description: Not found
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson with full content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    ModuleWithLessons:
      type: object
      required: [id, title, slug, description, difficulty, category, order, lessons]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        category:
          type: string
        order:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
    Lesson:
      type: object
      required: [id, moduleId, title, slug, order, duration]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        slug:
          type: string
        order:
          type: integer
        duration:
          type: integer
        completed:
          type: boolean
    LessonDetail:
      type: object
      required: [id, moduleId, title, slug, order, duration, content, commands]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        slug:
          type: string
        order:
          type: integer
        duration:
          type: integer
        content:
          type: string
        commands:
          type: array
          items:
            $ref: "#/components/schemas/Command"
        completed:
          type: boolean
        nextLessonId:
          type: ["integer", "null"]
        prevLessonId:
          type: ["integer", "null"]
    Command:
      type: object
      required: [command, description, output]
      properties:
        command:
          type: string
        description:
          type: string
        output:
          type: string
        dangerous:
          type: boolean
...