# lib/api-zod/src/generated/types/command.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kali Academy cybersecurity learning platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Command {
  command: string;
  description: string;
  output: string;
  dangerous?: boolean;
}