# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kali Academy cybersecurity learning platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  moduleId: number;
  title: string;
  slug: string;
  order: number;
  duration: number;
  completed?: boolean;
}