# lib/api-zod/src/generated/types/moduleWithLessons.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kali Academy cybersecurity learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from './lesson';
import type { ModuleWithLessonsDifficulty } from './moduleWithLessonsDifficulty';

export interface ModuleWithLessons {
  id: number;
  title: string;
  slug: string;
  description: string;
  difficulty: ModuleWithLessonsDifficulty;
  category: string;
  order: number;
  lessons: Lesson[];
}